Refactor : game controller, domain, repository, dto, service 디렉토리 변경 auth -> game
